move helpers.redraw.js to helpers/redraw.js, browserpass/browserpass-extension#320
